+++ src/main/java/com/bluehair/hanghaefinalproject/common/response/error/ErrorCode.java
@@ -1,4 +1,4 @@
-package com.bluehair.hanghaefinalproject.common.success;
+package com.bluehair.hanghaefinalproject.common.response.error;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,11 +7,11 @@ import org.springframework.http.HttpStatus;
 // Domain이 커지게 될 경우, 해당 객체를 Abstract로 선언 후 Domain별 extend
 @Getter
 @AllArgsConstructor
-public enum SucessCode {
-    // Member
-    SIGNUP_MEMBER(HttpStatus.OK, "회원 가입 성공", 2000);
+public enum ErrorCode {
+    INVALID_EMAIL(HttpStatus.OK, "유효하지 않은 이메일입니다.", 4041);
 
     private final HttpStatus httpStatus;
     private final String message;
     private final Integer customHttpStatusCode;
+
 }
