+++ src/main/java/com/bluehair/hanghaefinalproject/common/response/success/SuccessResponse.java
@@ -1,4 +1,4 @@
-package com.bluehair.hanghaefinalproject.common.success;
+package com.bluehair.hanghaefinalproject.common.response.success;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.Builder;
@@ -8,15 +8,15 @@ import org.springframework.http.ResponseEntity;
 @Getter
 @Builder
 @JsonInclude(JsonInclude.Include.NON_NULL)
-public class Response<T> {
+public class SuccessResponse<T> {
     private final Integer customHttpStatus;
     private final String message;
     private T data;
 
-    public static<T> ResponseEntity<Response<T>> toResponseEntity(SucessCode sucessCode, T data){
+    public static<T> ResponseEntity<SuccessResponse<T>> toResponseEntity(SucessCode sucessCode, T data){
         return ResponseEntity
                 .status(sucessCode.getHttpStatus())
-                .body(Response.<T>builder()
+                .body(SuccessResponse.<T>builder()
                         .customHttpStatus(sucessCode.getCustomHttpStatusCode())
                         .message(sucessCode.getMessage())
                         .data(data)
